+++ src/sll/parse.c
@@ -406,6 +406,9 @@ static void _read_object_internal(sll_file_t* rf,sll_source_file_t* sf,sll_read_
 				else if (*str==':'&&*(str+1)=='?'){
 					o->t=SLL_NODE_TYPE_TYPEOF;
 				}
+				else if (*str=='.'&&*(str+1)=='?'){
+					o->t=SLL_NODE_TYPE_NAMEOF;
+				}
 				else if (*str=='&'&&*(str+1)==':'){
 					o->t=SLL_NODE_TYPE_DECL;
 				}
@@ -462,6 +465,9 @@ static void _read_object_internal(sll_file_t* rf,sll_source_file_t* sf,sll_read_
 				}
 				else if (*str=='!'&&*(str+1)=='='&&*(str+2)=='='){
 					o->t=SLL_NODE_TYPE_STRICT_NOT_EQUAL;
+				}
+				else if (*str=='&'&&*(str+1)==':'&&*(str+2)=='?'){
+					o->t=SLL_NODE_TYPE_NAMEOF_TYPE;
 				}
 				else if (*str=='<'&&*(str+1)=='<'&&*(str+2)=='<'){
 					o->t=SLL_NODE_TYPE_CONTINUE;